test: split up group and channel tests
When developing, this makes you not have to wait
for the group tests to finish before the channel tests start.

This was an easy split, only had to adjust imports
and copy-paste the `beforeAll` and similar stuff.
